Edit reply function on redirect.py to mention user
This will generate a change to avoid confusion. As it currently stands, the bot removes the original message. This is by design.

In order to avoid context or confusion the bot will no longer just paste the URL and instead will mention whomever sent the original message, as well as modify the URL accordingly.